backend/src/requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
sqlalchemy==2.0.21
asyncpg==0.28.0
psycopg2-binary==2.9.8
alembic==1.12.0
python-dotenv==1.0.0
aiogram==3.1.1